chore: lots of cleanup in preparation for a node.js entry point
Move common files and tests into `common` folders.
Rename test util to `test-util` for clarity.
Remove unused `brick` test.
