+++ lib/common/utils.ts
