+++ scripts/sauce/sauce_connect_setup.sh
@@ -12,7 +12,7 @@ set -e -o pipefail
 # before_script:
 #   - curl https://gist.github.com/santiycr/5139565/raw/sauce_connect_setup.sh | bash
 
-CONNECT_URL="https://saucelabs.com/downloads/sc-4.3.13-linux.tar.gz"
+CONNECT_URL="https://saucelabs.com/downloads/sc-4.3.14-linux.tar.gz"
 CONNECT_DIR="/tmp/sauce-connect-$RANDOM"
 CONNECT_DOWNLOAD="sc-latest-linux.tar.gz"
