--- test/integration/brick.spec.ts
@@ -1,4 +1,0 @@
-// brick does some funny stuff with property descriptors
-// see:
-//
-// this file is intentionally empty; if `lib/brick.js` doesn't error, then we're cool
