+++ test/test-env-setup.ts
@@ -1,5 +1,2 @@
 (<any>jasmine).DEFAULT_TIMEOUT_INTERVAL = 2000;
 import '../lib/jasmine/jasmine';
-import './util';
-
-
